Write to separate log files (by process id):
This brings in changes made to utils where we will now separate
log files by process.

Writing to the same log file from multiple python processes is
unsupported.

This will ensure that we create a separate log file for celery
and the app and thereby ensure we do not lose any log lines.