restart celery between every task
we've been seeing issues with large amounts of memory not being
released from the zip_and_send_letter_pdfs task. While we'd like to
solve this issue, for now, since the ftp app is extremely long lived,
we'll restart the processes, clearing out the memory.

This requires us to do some wrangling. Due to reasons I don't quite
understand, the first worker process is started with the flask context
active (this happens on the final line of `run_celery.py`, when we run
`application.app_context().push`). However, the second process (after
the previous process was killed) doesn't have access to the current
app through `flask.current_app`. It feels like the import chain hasn't
run but somehow it's still got a hold of the celery object? My
suspicion is it's due to us using init_app rather than __init__, but
not sure.

To solve this, have the NotifyTask be a private class inside init_app
so that it gets access to the app object.