don't upload to dvla if file already in zips_sent
we can't rely on checking if the file is already present on dvla's ftp
server to see if we've already uploaded it - dvla might have already
deleted the file by the time the duplicate task runs. We upload a
record of the files we uploaded to the zips_sent subfolder in our s3
bucket as soon as we finish the ftp transfer, so we can check to see if
that file exists before we try commencing another ftp, in case we're in
a second duplicate the ftp transfer, so we can check to see if
that file exists before we try commencing another ftp, in case we're in
a second duplicate task

this isn't foolproof (a second task could run the check just before the
first task uploads the zips_sent), however in that case it's likely
that the ftp will fail as the file should still be on the remote
server if it's within seconds.