replace pysftp with the sftp in a subprocess controlled by pexpect
for reasons unknown, pysftp is slow. Quadratically slow, replicatable
locally. It might be related to https://github.com/paramiko/paramiko/pull/1479.

While waiting for a patch to be released, here's us invoking sftp
directly (controlling it via the pexepct library).

Note: I'd much rather wait for paramiko to be patched. This is a
seriously dangerous PR that I'd consider WIP, it's highly brittle and
the expect library (which controls interactive prompts programatically)
is quite terrifying.