Create different target to upload paas artifact
The jenkins job is calling `upload-codedeploy-artifact` 3 times in a row
in order to upload the zip file to each respective environment.

Given that the jenkins bucket is located in the tools environment, this
approach meant that a) we don't have access to that bucket and b) even
if we did we would be uploading the file 3 times.

To avoid both of these issues, `upload-paas-artifact` becomes a
target on its own.