Use default boto instead of our forked version
Note, it's the addition of `[sqs]` to the celery requirement that brings
in the `boto` dependancy in place of the fork we have removed.

We weren't sure what was stopping us making this change but have tested
locally this app can connect to sqs and read messages from a queue.