Remove unneeded configs that are now default
https://docs.celeryproject.org/en/v4.1.0/userguide/configuration.html#accept-content
accept_content default is json as per docs so we can remove this

https://docs.celeryproject.org/en/v4.1.0/userguide/configuration.html#accept-content
task_serializer default is json as per docs so we can remove this. Note,
this changed in version 4

https://docs.celeryproject.org/en/v4.1.0/userguide/configuration.html#accept-content
enable_utc is enabled by default as per docs so we can remove this. This
follows the similar lack of this setting in the antivirus and template
preview celery configs