install celery sqs dependencies ourselves
as suggested by https://github.com/celery/celery/pull/6536#issuecomment-750270731

installing celery deps ourselves means we can bump pycurl to v7.43.0.6,
This means we need to make sure we've installed celery's sqs
dependencies. At the moment this is just boto3 and pycurl.

we're bumping pycurl to make it easier to install on MacOS
https://github.com/celery/celery/pull/6536#issuecomment-750270731