pin boto3 to v1.16
we don't currently pin boto3. Probably good to ensure we're using an up
to date version of the library. As the patch version updates almost
daily, and almost always just bumps bits of aws that we don't use,
we don't think there's a need to pin to a specific patch, but just keep
major version up to date.