add requirements_dev.txt for local development
on MacOS, pycurl needs to be installed with the flag `--with-openssl`.
However, setting that as a --global-option in pip causes the entire pip
command to attempt to install all packages from source rather than from
wheels.

We need to install everything else from wheels, so move pycurl into its
own requirements file that we'll only use for running locally. (We don't
think this openssl flag is required on concourse/paas).