Extend long polling time to 20 seconds
This makes it consistent with antivirus and template preview

Default is 10 seconds as per
https://docs.celeryproject.org/en/v4.4.7/getting-started/brokers/sqs.html#long-polling

Note, I understand this to mean that celery will start polling, wait up
to 20 seconds to find all the tasks to grab, will grab them and then
wait another second (the default polling_interval) before trying again.
This difference between 10 and 20 seconds isn't too different, it's not
the end of the world if we wait a little longer to pick tasks up for the
FTP app. However, it should save some money as most of the time the FTP
app isn't going to find any tasks. It also makes things consistent which
is another reason to do this.