Make mock_remote_file fail on exceptions
The `__exit__` function is responsible for handling the exceptions in the
`with` statement.

According to step 7 here: https://docs.python.org/3/reference/compound_stmts.html#the-with-statement

If `__exit__` is called due to an exception and the return value is
True, the exception is suppressed.
If the return value is False, the exception is reraised.
If `__exit__` was _not_ called because of an exception, the return value
is ignored.

This means that we can have a function that always returns False and
will only matter in case of exceptions