Add tighter measure and logging of operations
The performance improvements in the previous PRs did not bring about the
same level of improvement as when we did our own testing.

The changes here make it easier to see the size of the file and the
duration of the upload operation _only_ , and also measure the overall duration
of the `upload_zip` function.

Include filename in the log lines