Use ThreadPoolExecutor to parallelise getting files from S3
This is in order to further improve performance of the FTP application.

Specifically we found that the `get_zip_of_letter_pdfs_from_s3` function
was responsible for the 33 of the 37 seconds of the total task duration.

This commit is using the python 3 feature that allows us to launch tasks
in parallel, using the ThreadPoolExecutor:

https://docs.python.org/3/library/concurrent.futures.html#concurrent.futures.Future