Update NotifyTask to match API
This copies the new version of the task we have in the API app, as
well as the tests for it. Using the new version:

- Standardises the logging and metrics for success / failure.
- Adds support for "request_id" (to build on in future PRs).

Although the tasks in this app are never invoked via a web request,
it's still possible the "request_id" kwarg may be set e.g. due a bug
in the API. It's prudent and consistent to handle its presence.

A couple of other notes about the changes here:

- Upgrading freezegun is necessary to freeze monotonic time.
- Moving celery.init_app ensures StatsD is initialised.