Add a retry for the zip_and_send_letter_pdfs task.
If the task fails to send the PDFs, retry the task 5 times with a delay interval of 5 minutes (300 seconds).
If the file is sent twice DVLA will not process the second file with the duplicate name, however, if the first file is corrupt they will then use the seconds file that was successful.
Only if the file fails to send after the fifth retry will the notifications be marked as a technical-failure.