Introduce freeze requirements using pip-tools
Not pinning all our requirements introduced a bug when we redeployed the
ftp app thinking that we had made no changes to dependencies.

We adapt the same approach as the other apps to pin our dependencies.

This follows the approach of the admin app which has started using
piptools rather than our previous 'hacky' script.

For some context see
https://github.com/alphagov/notifications-admin/pull/3753/files
and
https://github.com/alphagov/notifications-admin

Note, we don't use `VIRTUALENV_ROOT` in this makefile, we just assume
that people are in a virtualenv rather than using a separate `venv`
directory.