Remove pycurl check in bootstrap task
This causes the task to fail on CI (where we're looking to use it
instead of CI having its own setup [1]).

The README has extensive guidance on installing pycurl for local
development, so I think it's OK to remove the check here. I can't
find any specific pycurl setup for CI, which also raises a question
of whether it's actually necessary - maybe it's just a Mac v. Linux
issue, so no additional setup is required on CI.