remove mock asserts from pytest raises blocks
`with pyrest.raises(Exc)` blocks simply catch an exception - and if an
exception wasn't thrown inside the block, they raise an
AssertionFailure.

However, they can't re-write the code inside the block. If there are two
lines of code inside the block (a function call, then a mock assert),
and that first function call raises an exception, the second line (the
mock assertion in this case) won't be executed.

This commit unindents the mock asserts to ensure that they are called
properly. Some of the asserts (on the mock retry) were called with the
wrong arguments.

However, the tests around MaxRetriesExceededError are still failing.
This is due to a bug in the task itself, and will be fixed in a separate
commit.