handle max retries correctly
Previously, we were catching, eg, a ClientError, and then calling
self.retry. This itself might throw a max retries exception. Exception
try catch blocks at the same level as that ClientError won't be called
for exceptions thrown from within the catch. Abstracted that handling of
max exceptions out to a separate function to keep it tidy.