Fix pip install failing on Mac M1 machines
Fixes this error:

        build/temp.macosx-12.2-arm64-3.9/_openssl.c:575:10: fatal error: 'openssl/opensslv.h' file not found
        #include <openssl/opensslv.h>
                 ^~~~~~~~~~~~~~~~~~~~

I did consider upgrading to the latest version, but since this is
a low-level dependency I think we should wait for the libraries
that depend on it to give the signal first.

Version 3.4+ introduce Rust into the build pipeline [^1], which I
recall we had problems with previously, but a test deploy indicates
there's nothing to worry about here.

I've also added jinja2-cli to support local deployments - in order
to generate the manifest file.

[^1]: https://github.com/pyca/cryptography/blob/main/CHANGELOG.rst#34---2021-02-07