Increase duration for creating and uploading zip
We set a seemingly-arbitrary timeout on this task for processing,
zipping and uploading letters to the DVLA FTP. Recently with a large NHS
campaign starting we've started to see timeouts on this task.

From https://github.com/alphagov/notifications-ftp/pull/364:

"I set the entire task timeout to 90 seconds and the sftp timeout to 30
seconds, which was a totally arbitrary decision that I am in no way
wedded to. I'm open to ideas. Normally tasks take up to 20 seconds."

Let's double it so that it has a bit more time to process letters.